Organize flask server into modules; This prepares splitting routes from main.py
