+++ server/lib/barchart_template.py
