+++ server/lib/coordinate_calculator.py
