+++ server/models/stat_config.proto
