+++ server/templates/base.html
