+++ server/templates/gni_base.html
