+++ server/templates/gni_scatter.html
