+++ server/templates/kg_chart.html
