+++ server/templates/mcf_playground.html
