+++ server/templates/translator.html
