+++ server/tests/main_test.py
@@ -1,3 +1,17 @@
+# Copyright 2020 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import pytest
 import unittest
 import logging
@@ -51,7 +65,7 @@ class TestApiParentPlaces(unittest.TestCase):
         else:
           return {}
 
-    @patch('datacommons.send_request')
+    @patch('services.datacommons.send_request')
     def test_api_parent_places(self, mock_send_request):
         mock_send_request.side_effect = self.side_effect
         response = app.test_client().get('/api/parent-place/geoId/0649670')
